+++ cpp_08/ex01/Span.h
@@ -6,7 +6,7 @@
 /*   By: eabdelha <eabdelha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/11 08:34:28 by eabdelha          #+#    #+#             */
-/*   Updated: 2022/08/27 15:24:26 by eabdelha         ###   ########.fr       */
+/*   Updated: 2022/08/29 14:47:27 by eabdelha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,21 +20,21 @@
 
 class Span
 {
-    typedef std::vector<int>::iterator iterator;
-    typedef std::vector<int>              type_name;
+    typedef std::vector<int>::iterator  iterator;
+    typedef std::vector<int>            type_name;
     
-    type_name numbers;
-    unsigned  size;
+    type_name     numbers;
+    unsigned int  size;
   
     public:
-        Span(unsigned n);
+        Span(unsigned int n);
         Span(const Span &span);
         Span &operator = (const Span &span);
-        int const &operator[](size_t index) const ;
-        void addNumber(size_t num);
+        int const &operator[](int index) const ;
+        void addNumber(int num);
         void addNumber(iterator begin, iterator end);
         size_t shortestSpan(void);
-        size_t longestSpan(void);
+        size_t longestSpan(void);   
         class noSpan : public std::exception {
           const char *what() const throw();  
         };
